% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{aplicacao_todas_as_colunas}
\alias{aplicacao_todas_as_colunas}
\title{aplicacao_todas_as_colunas}
\usage{
aplicacao_todas_as_colunas(colunaBase, tabela, funcao, valoresPadrao = NULL)
}
\arguments{
\item{colunaBase}{uma variavel de fatores de um dataframe}

\item{tabela}{um dataframe sobre o qual serao realizadas as operacoes}

\item{funcao}{a funcao que sera aplicada sobre as colunas de \code{tabela}}

\item{valoresPadrao}{o conjunto de valores que sera considerado para a aplicacao da funcao}
}
\value{
retorna os resultados da aplicacao da funcao em cada coluna
}
\description{
funcao em desenvolvimento que aplica uma funcao a todas as colunas de uma tabela de acordo com uma coluna referencia
}
