% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{aplicador_tabelas}
\alias{aplicador_tabelas}
\title{aplicador_tabelas
funcao ainda em desenvolvimento que aplica uma funcao a todas as variaveis de uma tabela}
\usage{
aplicador_tabelas(
  tabela,
  funcao,
  funcao2 = NULL,
  valoresPadrao = NULL,
  colunaBase = NULL
)
}
\arguments{
\item{tabela}{o dataframe sobre o qual sera realizada a operacao}

\item{funcao}{em aplicacao}

\item{funcao2}{em aplicacao}
}
\description{
aplicador_tabelas
funcao ainda em desenvolvimento que aplica uma funcao a todas as variaveis de uma tabela
}
