% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{autor}
\alias{autor}
\title{variavel de autor
 @description 
 seta o autor que esta escrevendo no momento
 @export}
\format{An object of class \code{character} of length 1.}
\usage{
autor
}
\description{
variavel de autor
 @description 
 seta o autor que esta escrevendo no momento
 @export
}
\keyword{datasets}
