% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{completitude}
\alias{completitude}
\title{completitude}
\usage{
completitude(variavel, log = TRUE)
}
\arguments{
\item{variavel}{um vetor, ou variavel (coluna) de um dataframe sobre a qual a operacao sera realizada}
}
\value{
retorna o calculo da completitude de \code{variavel}
}
\description{
completitude
}
\examples{
completitude(dengue2013$SEM_NOT)
}
