% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{completitude_variaveis_de_uma_tabela}
\alias{completitude_variaveis_de_uma_tabela}
\title{completitude_variaveis_de_uma_tabela}
\usage{
completitude_variaveis_de_uma_tabela(tabela, registrar = FALSE, log = TRUE)
}
\arguments{
\item{tabela}{um dataframe sobre o qual sera realizada a operacao}

\item{registrar}{indica se o resultado final deve ser registrado em um documento no formato xlsx}
}
\value{
retorna um dataframe com a completitude de todas as variaveis de uma tabela
}
\description{
Calcula a completitude de todas as variaveis (colunas) de um dataframe
}
\examples{
completitude_variaveis_de_uma_tabela(tabela = dengue2013)
completitude_variaveis_de_uma_tabela(tabela = dengue2013, registrar = TRUE)
}
