% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{completitudeRelacionada}
\alias{completitudeRelacionada}
\title{completitude_relacionada}
\usage{
completitudeRelacionada(
  tabela,
  variaveis_de_referencia,
  variavel_para_avaliacao,
  valoresPadrao = NULL,
  nota = NULL
)
}
\arguments{
\item{tabela}{o dataframe em que os calculos serao realizados}

\item{variaveis_de_referencia}{o conjunto de variaveis utilizadas como referencia, quando combinadas devem indicar uma situacao}

\item{variavel_para_avaliacao}{a variavel sobre a qual sera calculada a completitude}

\item{valoresPadrao}{um conjunto de vetores, com os fatores de cada uma das variaveis de referencia, respectivamente}
}
\value{
Retorna a completitude da \code{variavel_para_avaliacao} contendo apenas as linhas que possuiam determinados conjuntos de fatores
nas \code{variaveis_de_referencia} de acordo com os \code{valoresPadrao}
}
\description{
calcula a completitude de uma variavel de acordo com um conjunto de fatores presentes em um conjunto de variaveis de referencia
}
\examples{
completitude_relacionada(dengue2013, ("CS_SEXO"), c("CS_GESTANT"), c(c("F")))
}
