% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{consistencia}
\alias{consistencia}
\title{consistencia}
\usage{
consistencia(
  colunaA = NULL,
  colunaB = NULL,
  coluna_adicional_parametro = NULL,
  data = TRUE,
  valoresPadrao = NULL,
  valores_padrao_adicionais = NULL,
  formato = NULL,
  nota = NULL
)
}
\arguments{
\item{colunaA}{variavel de um dataframe que sera utilizada como referencia}

\item{colunaB}{variavel de um dataframe que sera checada}

\item{coluna_adicional_parametro}{variavel de um dataframe que sera utilizada para filtrar a tabela}

\item{data}{indica se a funcao sera aplicada sobre variaveis de data}

\item{valoresPadrao}{o conjunto de fatores a serem filtrados em \code{colunaA}}

\item{valores_padrao_adicionais}{o conjunto de fatores a serem filtrador em \code{coluna_adicional_parametro}}

\item{formato}{o formato da data utilizado}
}
\value{
retorna a consistencia com relacao as variaveis determinadas
}
\description{
calcula a consistencia de variaveis de fatores e datas
}
\examples{
consistencia(colunaA = dengue2013$DT_SORO, colunaB = dengue2013$RESUL_SORO, duas_colunas_data = FALSE)
consistencia(colunaA = dengue2013$RESUL_SORO, colunaB = dengue2013$DT_SORO, duas_colunas_data = FALSE)
consistencia(colunaA = dengue2013$RESUL_SORO, colunaB = dengue2013$DT_SORO, valoresPadrao = c("2", "4"), duas_colunas_data = FALSE)
consistencia(colunaA = dengue2013$RESUL_SORO, colunaB = dengue2013$DT_SORO, coluna_adicional_parametro = dengue2013$NU_ANO, duas_colunas_data = FALSE)
consistencia(colunaA = dengue2013$RESUL_SORO, colunaB = dengue2013$DT_SORO,coluna_adicional_parametro = dengue2013$NU_ANO,
             valoresPadrao = c("2", "4"), valores_padrao_adicionais =  c("2013"),duas_colunas_data = FALSE)
}
