% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{definirAnalista}
\alias{definirAnalista}
\title{define o autor para os logs}
\usage{
definirAnalista(nome)
}
\arguments{
\item{nome}{o nome do ator que estah trabalhando com o pacote}
}
\description{
define o autor para os logs
}
