% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{definirAutor}
\alias{definirAutor}
\title{define o autor para os logs}
\usage{
definirAutor(nome)
}
\arguments{
\item{nome}{o nome do ator que estah trabalhando com o pacote}
}
\description{
define o autor para os logs
}
