% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{definirParagrafo}
\alias{definirParagrafo}
\title{define paragrafo a ser escrito}
\usage{
definirParagrafo(texto)
}
\arguments{
\item{paragrafo}{o paragrafo que serah escrito antes das funcoes registradas no relatorio rmd}
}
\description{
define paragrafo a ser escrito
}
