% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{dengue2013}
\alias{dengue2013}
\title{dengue2013}
\format{Um data frame com 788 linhas e 43 variaveis:}
\source{
MS (Ministério da Saúde)
}
\usage{
dengue2013
}
\description{
dados coletados acerca da situacao epidemiologica brasileira acerca dos dados de dengue
}
\keyword{datasets}
