% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{duracao_em_dias}
\alias{duracao_em_dias}
\title{duracao_em_dias}
\usage{
duracao_em_dias(
  coluna_data_x = NULL,
  coluna_data_y = NULL,
  mediana = FALSE,
  quantidade_dias_referencia = NULL,
  variavelIdentificador = NULL,
  valoresPadrao = NULL,
  formato = NULL,
  nota = NULL
)
}
\arguments{
\item{coluna_data_x}{vetor de um dataframe com a data final do evento}

\item{coluna_data_y}{vetor de um dataframe com a data de inicio do evento}

\item{mediana}{indica se a mediana devera ser calculada}

\item{quantidade_dias_referencia}{em implementacao}

\item{variavelIdentificador}{uma variavel qualitativa do datafram sobre o qual serao determinados os eventos de interesse (quando preenchido)}

\item{valoresPadrao}{os casos que devem ser considerados para \code{variavelIdentificador}}

\item{formato}{o formato de data que deve ser utilizado}
}
\value{
prints com sumarios da distribuicao da duracao do evento, media e mediana da duracao do evento
}
\description{
funcao ainda em desenvolvimento, tera como objetivo o calculo da duracao media em dias de dois eventos, recebendo a data inicio x e a data final x do evento
e outras informacoes
}
\examples{
duracao_em_dias(dengue2013$DT_NOTIFIC, dengue2013$DT_SIN_PRI, mediana = TRUE
               ,valoresPadrao = c("2", "4"), variavel_identificador = dengue_2013$RESUL_SORO)
}
