% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{extratificacaoIdade}
\alias{extratificacaoIdade}
\title{extratificacaoIdade
 Constroi extratos para datas (ano de nascimento) ou valores inteiros}
\usage{
extratificacaoIdade(
  tabela,
  coluna,
  extratos = NULL,
  nomesExtratos = NULL,
  nomeNovaColuna = NULL,
  coluna_eh_data_nascimento = FALSE,
  formatoData = NULL
)
}
\arguments{
\item{tabela}{o dataframe em uso}

\item{coluna}{a coluna que sera utilizada}

\item{extratos}{a definicao dos extratos nos quais havera a divisao}

\item{nomesExtratos}{os nomes dos extratos definidos}

\item{nomeNovaColuna}{nome da coluna que sera adicionada ao dataframe com os novos extratos}

\item{coluna_eh_data_nascimento}{informa se a coluna apresenta informacoes de data de nascimento}

\item{formatoData}{informa o formato em que a data esta construida}
}
\description{
extratificacaoIdade
 Constroi extratos para datas (ano de nascimento) ou valores inteiros
}
