% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{filtro_mes_variavel}
\alias{filtro_mes_variavel}
\title{filtro_mes_variavel
funcao experimental que filtra por mes e por variavel algumas tabelas para a construcao de um indicador epidemiologico}
\usage{
filtro_mes_variavel(
  tabelaA,
  coluna_quantidade_somatorioA,
  tabelaB = NULL,
  coluna_quantidade_somatorioB = NULL,
  tabelaC = NULL,
  coluna_quantidade_somatorioC = NULL,
  colunafiltroA,
  colunafiltroB,
  colunafiltroC,
  colunaDataA,
  colunaDataB,
  colunaDataC = NULL,
  meses,
  anos,
  indicador,
  valores_padrao_aplicados = NULL,
  casos = NULL,
  valoresPadrao = NULL
)
}
\description{
filtro_mes_variavel
funcao experimental que filtra por mes e por variavel algumas tabelas para a construcao de um indicador epidemiologico
}
