% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{porcentagem_verdadeiro_positivo}
\alias{porcentagem_verdadeiro_positivo}
\title{porcentagem_verdadeiro_positivo}
\usage{
porcentagem_verdadeiro_positivo(variavel, chavePositivo, chave_falso_positivo)
}
\arguments{
\item{variavel}{a variavel de um dataframe com os fatores investigados}

\item{chave_falso_positivo}{os fatores indicando quais sao os fatores que representam casos aferidos erroneamente como positivos}

\item{chavePositvo}{os fatores indicando quais sao os fatores verdadeiramente positivos}
}
\value{
retorna a porcentagem dos valores positivos que eram verdadeiros
}
\description{
porcentagem_verdadeiro_positivo
}
\examples{
porcentagem_verdadeiro_positivo(dengue2013$RESUL_SORO, "2", "4")
}
