% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{quantidade_de_observacoes}
\alias{quantidade_de_observacoes}
\title{quantidade_de_observacoes}
\usage{
quantidade_de_observacoes(variavel, valoresPadrao = NULL)
}
\arguments{
\item{variavel}{recebe um vetor de fatores}

\item{valoresPadrao}{o conjunto de fatores que sera considerado para realizar a operacao}
}
\value{
retorna a quantidade de cada fator de \code{valoresPadrao} em \code{variavel}
}
\description{
calcula com forca bruta
}
\examples{
quantidade_de_observacoes(variavel = dengue2013$MUNICIPIO)
quantidade_de_observacoes(variavel = dengue2013$RESUL_SORO, valoresPadrao = c("2", "4"))
}
