% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{representatividade}
\alias{representatividade}
\title{representatividade}
\usage{
representatividade(variavel, valoresPadrao)
}
\arguments{
\item{variavel}{a variavel de um dataframe que contenha valores categoricos (fatores)}

\item{valoresPadrao}{o conjunto de fatores sobre os quais sera calculada a representatividade}
}
\description{
calcula a representatividade de um conjunto de fatores em uma variavel
}
\examples{
representatividade(dengue2013$RESUL_SORO, c("2","4"))
}
