% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{representatividadeCondicionada}
\alias{representatividadeCondicionada}
\title{representatividadeCondicionada}
\usage{
representatividadeCondicionada(
  colunaFiltro,
  valoresFiltro,
  colunaAlvo,
  valoresPadrao
)
}
\arguments{
\item{colunaFiltro}{recebe a coluna de um dataframe que registra os fatores}

\item{valoresFiltro}{recebe os fatores da variavel \code{colunaFiltro} que serao utilizados}

\item{colunaAlvo}{recebe a coluna do mesmo dataframe que \code{colunaFiltro} do qual sera calculada a representatividade}

\item{valoresPadrao}{o conjunto de fatores que irao ser calculados perante o total de observacoes}
}
\value{
Retorna a porcentagem dos valores de \code{valoresPadrao} encontrados na \code{colunaAlvo}, que possuem tambem os \code{valoresFiltro} na \code{colunaFiltro}
}
\description{
funcao que calcula a representavidade de uma variavel, de acordo com uma variavel referencia de fatores (para um conjunto de fatores)
}
\examples{
representatividadeCondicionada(colunaFiltro = dengue2013$RESUL_SORO, 
                               valoresFiltro = c("2","4"), 
                               colunaAlvo = dengue2013$NU_ANO, 
                               valoresPadrao = c("2013")
                               )
}
