% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{tabela_variaveis_avaliacao}
\alias{tabela_variaveis_avaliacao}
\title{tabela_variaveis_avaliacao}
\usage{
tabela_variaveis_avaliacao(
  variavelBase,
  variavel,
  funcao,
  valoresPadrao = NULL
)
}
\arguments{
\item{variavelBase}{variavel referencia de um dataframe}

\item{variavel}{variavel de um dataframe sobre a qual serao realizadas as operacoes}

\item{funcao}{funcao que sera aplicada sobre a \code{variavel}}

\item{valoresPadrao}{o conjunto de fatores considerado nos quais serao calculadas a operacao}
}
\value{
retorna o resultado e registra em tabela xlsx da \code{funcao} sobre a \code{variavel} aplicado apenas as linhas que possuem os \code{valoresPadrao} na \code{variavelBase}
}
\description{
tabela_variaveis_avaliacao
}
\examples{
tabela_variaveis_avaliacao(dengue2013$CS_SEXO, dengue2013$RESUL_SORO, representatividade, valoresPadrao = c("2"))
}
