% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{validade}
\alias{validade}
\title{validade}
\usage{
validade(
  variavel,
  valoresPadrao,
  intervalo = FALSE,
  data = FALSE,
  formatoData = NULL,
  nota = NULL
)
}
\arguments{
\item{variavel}{um vetor, normalmente proveniente de uma variavel de um dataframe, ao qual serao realizadas as operacoes}

\item{valoresPadrao}{indica todos os valores que a \code{variavel} de fatores sobre a qual sera realizada a operacao deve conter}

\item{data}{indica se a \code{variavel} de fatores possui o formato de data}

\item{formatoData}{indica qual formato de data a funcao deve considerar, so pode ser utilizada se \code{data} for verdadeiro}
}
\value{
retorna a validade de \code{variavel} em character
}
\description{
Calcula a validade para um vetor
}
\examples{
validade(variavel = dengue2013$RESUL_SORO, valoresPadrao = c("1", "2", "3","4"))
validade(variavel = dengue2013$DT_NOTIFIC, valoresPadrao = c("2013/01/01","2013/12/01"), data = TRUE)
-O valor a esquerda em valores padrao eh o limite inferior (por exemplo se quiser incluir 1000, coloque 999)
-O valor a direita em valores padrao eh o limite superior (por exemplo, se quiser incluir 3000, coloque 3001)
validade(dengue_2013$NU_IDADE_N, valoresPadrao = c("1000", "3000"), intervalo = TRUE) 
}
