% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{valores_em_branco_para_NA}
\alias{valores_em_branco_para_NA}
\title{valores_em_branco_para_NA}
\usage{
valores_em_branco_para_NA(variavel, nota = NULL)
}
\arguments{
\item{um}{vetor sobre o qual sera realizada a operacao}
}
\value{
retorna o vetor, com todas as observacoes que estavam em branco trocados por NA
}
\description{
valores_em_branco_para_NA
}
\examples{
valores_em_branco_para_NA(dengue2013$DT_NOTIFIC)
}
