% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{verificaNas}
\alias{verificaNas}
\title{verificaNas}
\usage{
verificaNas(tabela, nota = NULL)
}
\arguments{
\item{tabela}{um dataframe sobre o qual serao realizadas as operacoes}
}
\value{
retorna um hash com o nome de cada coluna da tabela e os valores em branco contidos nessa coluna
}
\description{
verificaNas
}
\examples{
verificaNas(dengue2013)
}
