% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{extratificacaoIdade}
\alias{extratificacaoIdade}
\title{extratificacaoIdade
 Constroi extratos para datas (ano de nascimento) ou valores inteiros}
\usage{
extratificacaoIdade(
  tabela,
  coluna,
  extratos = NULL,
  nomesExtratos = NULL,
  nomeNovaColuna = NULL,
  coluna_eh_data_nascimento = FALSE,
  formatoData = NULL
)
}
\arguments{
\item{tabela}{o dataframe ao qual serah adicionada a nova coluna}

\item{coluna}{a coluna que sera extratificada (podendo ser numeric ou datas de nascimento)}

\item{extratos}{a definicao dos extratos nos quais havera a divisao}

\item{nomesExtratos}{os nomes dos extratos definidos}

\item{nomeNovaColuna}{nome da coluna que sera adicionada ao dataframe com os novos extratos}

\item{coluna_eh_data_nascimento}{informa se a coluna apresenta informacoes de data de nascimento}

\item{formatoData}{informa o formato em que a data está respresentada. Obs: O parâmetro só pode ser preenchido se coluna_eh_data_nascimento for verdadeiro.}
}
\value{
O dataframe \code{tabela} com uma nova variável representando a \code{coluna} extratificada com os \code{extratos} definidos, ou padrão.
}
\description{
extratificacaoIdade
 Constroi extratos para datas (ano de nascimento) ou valores inteiros
}
