% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{extratificacaoIdade}
\alias{extratificacaoIdade}
\title{extratificacaoIdade}
\usage{
extratificacaoIdade(
  tabela,
  coluna,
  extratos = NULL,
  nomesExtratos = NULL,
  nomeNovaColuna = NULL,
  coluna_eh_data_nascimento = FALSE,
  formatoData = NULL,
  nota = NULL
)
}
\arguments{
\item{tabela}{o dataframe ao qual serah adicionada a nova coluna}

\item{coluna}{a coluna que sera extratificada (podendo ser numeric ou datas de nascimento)}

\item{extratos}{a definicao dos extratos nos quais havera a divisao}

\item{nomesExtratos}{os nomes dos extratos definidos}

\item{nomeNovaColuna}{nome da coluna que sera adicionada ao dataframe com os novos extratos}

\item{coluna_eh_data_nascimento}{informa se a coluna apresenta informacoes de data de nascimento}

\item{formatoData}{informa o formato em que a data esta respresentada. Obs: O parametro só pode ser preenchido se coluna_eh_data_nascimento for verdadeiro.}
}
\value{
O dataframe \code{tabela} com uma nova variavel representando a \code{coluna} extratificada com os \code{extratos} definidos, ou padrao.
}
\description{
Constroi extratos para datas (ano de nascimento) ou valores inteiros
}
\examples{
extratificacaoIdade(tabela = dengue2013,
                    coluna = dengue2013$DT_NASC,
                    coluna_eh_data_nascimento = TRUE,
                    nomesExtratos = c("<1", "1-4", "5-9", "10-19", "20-29", "30-29", "40-49", "50-59", "60-69", "70+")
)
}
